Implement simple upcall bridge
