+++ src/main/java/de/tum/www1/orion/bridge/core/ArtemisCoreUpcallBridge.java
@@ -1,7 +1,8 @@
-package de.tum.www1.orion.bridge;
+package de.tum.www1.orion.bridge.core;
 
 import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.project.Project;
+import de.tum.www1.orion.bridge.JavaUpcallBridge;
 import org.jetbrains.annotations.NotNull;
 
 public interface ArtemisCoreUpcallBridge extends JavaUpcallBridge {
